Fix private
